chore: format

+++ src/pyth/external/IPyth.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity ^0.8.28;
 
-import "./PythStructs.sol";
-import "./IPythEvents.sol";
+import {PythStructs} from "./PythStructs.sol";
+import {IPythEvents} from "./IPythEvents.sol";
 
 /// @title Consume prices from the Pyth Network (https://pyth.network/).
 /// @dev Please refer to the guidance at https://docs.pyth.network/documentation/pythnet-price-feeds/best-practices for how to consume prices safely.
@@ -23,7 +23,7 @@ interface IPyth is IPythEvents {
     /// applications that require a sufficiently-recent price. Reverts if the price wasn't updated sufficiently
     /// recently.
     /// @return price - please read the documentation of PythStructs.Price to understand how to use this safely.
-    function getPriceNoOlderThan(bytes32 id, uint age) external view returns (PythStructs.Price memory price);
+    function getPriceNoOlderThan(bytes32 id, uint256 age) external view returns (PythStructs.Price memory price);
 
     /// @notice Returns the exponentially-weighted moving average price of a price feed without any sanity checks.
     /// @dev This function returns the same price as `getEmaPrice` in the case where the price is available.
@@ -44,7 +44,7 @@ interface IPyth is IPythEvents {
     /// applications that require a sufficiently-recent price. Reverts if the price wasn't updated sufficiently
     /// recently.
     /// @return price - please read the documentation of PythStructs.Price to understand how to use this safely.
-    function getEmaPriceNoOlderThan(bytes32 id, uint age) external view returns (PythStructs.Price memory price);
+    function getEmaPriceNoOlderThan(bytes32 id, uint256 age) external view returns (PythStructs.Price memory price);
 
     /// @notice Update price feeds with given update messages.
     /// This method requires the caller to pay a fee in wei; the required fee can be computed by calling
@@ -80,12 +80,12 @@ interface IPyth is IPythEvents {
     /// @notice Returns the required fee to update an array of price updates.
     /// @param updateData Array of price update data.
     /// @return feeAmount The required fee in Wei.
-    function getUpdateFee(bytes[] calldata updateData) external view returns (uint feeAmount);
+    function getUpdateFee(bytes[] calldata updateData) external view returns (uint256 feeAmount);
 
     /// @notice Returns the required fee to update a TWAP price.
     /// @param updateData Array of price update data.
     /// @return feeAmount The required fee in Wei.
-    function getTwapUpdateFee(bytes[] calldata updateData) external view returns (uint feeAmount);
+    function getTwapUpdateFee(bytes[] calldata updateData) external view returns (uint256 feeAmount);
 
     /// @notice Parse `updateData` and return price feeds of the given `priceIds` if they are all published
     /// within `minPublishTime` and `maxPublishTime`.
@@ -160,10 +160,10 @@ interface IPyth is IPythEvents {
     /// @param updateData Array containing exactly two price updates (start and end points for TWAP calculation)
     /// @param priceIds Array of price ids to calculate TWAP for
     /// @return twapPriceFeeds Array of TWAP price feeds corresponding to the given `priceIds` (with the same order)
-    function parseTwapPriceFeedUpdates(
-        bytes[] calldata updateData,
-        bytes32[] calldata priceIds
-    ) external payable returns (PythStructs.TwapPriceFeed[] memory twapPriceFeeds);
+    function parseTwapPriceFeedUpdates(bytes[] calldata updateData, bytes32[] calldata priceIds)
+        external
+        payable
+        returns (PythStructs.TwapPriceFeed[] memory twapPriceFeeds);
 
     /// @notice Similar to `parsePriceFeedUpdates` but ensures the updates returned are
     /// the first updates published in minPublishTime. That is, if there are multiple updates for a given timestamp,
